Lexical: Merged list nodes
