+++ resources/js/wysiwyg/lexical/list/__tests__/unit/LexicalListItemNode.test.ts
@@ -1265,99 +1265,5 @@ describe('LexicalListItemNode tests', () => {
         expect($isListItemNode(listItemNode)).toBe(true);
       });
     });
-
-    describe('ListItemNode.setIndent()', () => {
-      let listNode: ListNode;
-      let listItemNode1: ListItemNode;
-      let listItemNode2: ListItemNode;
-
-      beforeEach(async () => {
-        const {editor} = testEnv;
-
-        await editor.update(() => {
-          const root = $getRoot();
-          listNode = new ListNode('bullet', 1);
-          listItemNode1 = new ListItemNode();
-
-          listItemNode2 = new ListItemNode();
-
-          root.append(listNode);
-          listNode.append(listItemNode1, listItemNode2);
-          listItemNode1.append(new TextNode('one'));
-          listItemNode2.append(new TextNode('two'));
-        });
-      });
-      it('indents and outdents list item', async () => {
-        const {editor} = testEnv;
-
-        await editor.update(() => {
-          listItemNode1.setIndent(3);
-        });
-
-        await editor.update(() => {
-          expect(listItemNode1.getIndent()).toBe(3);
-        });
-
-        expectHtmlToBeEqual(
-          editor.getRootElement()!.innerHTML,
-          html`
-            <ul>
-              <li value="1">
-                <ul>
-                  <li value="1">
-                    <ul>
-                      <li value="1">
-                        <ul>
-                          <li value="1">
-                            <span data-lexical-text="true">one</span>
-                          </li>
-                        </ul>
-                      </li>
-                    </ul>
-                  </li>
-                </ul>
-              </li>
-              <li value="1">
-                <span data-lexical-text="true">two</span>
-              </li>
-            </ul>
-          `,
-        );
-
-        await editor.update(() => {
-          listItemNode1.setIndent(0);
-        });
-
-        await editor.update(() => {
-          expect(listItemNode1.getIndent()).toBe(0);
-        });
-
-        expectHtmlToBeEqual(
-          editor.getRootElement()!.innerHTML,
-          html`
-            <ul>
-              <li value="1">
-                <span data-lexical-text="true">one</span>
-              </li>
-              <li value="2">
-                <span data-lexical-text="true">two</span>
-              </li>
-            </ul>
-          `,
-        );
-      });
-
-      it('handles fractional indent values', async () => {
-        const {editor} = testEnv;
-
-        await editor.update(() => {
-          listItemNode1.setIndent(0.5);
-        });
-
-        await editor.update(() => {
-          expect(listItemNode1.getIndent()).toBe(0);
-        });
-      });
-    });
   });
 });
